Launch files init
I've created a new 'bringup' package for putting main system-spanning launch and config files. In it is a 'main_gokart_launch' and 'main_unity_launch' file. Each are used in the ros2 launch command for quickly launching nodes for the gokart and unity sim respectively.

An additional parameter is provided to allow disabling the visualisation nodes.
eg:
ros2 launch bringup main_gokart_launch vis:=False

This is currently a WIP the package.xml does not have dependencies specified and the package as a whole needs testing.